Convolutional Neural Network
This is the culmination of many months of work. I have amalgamated code to build a CNN, run a forward pass, a backwards pass, and optimizations in a training loop for this project. This project works, but is still in early development for now. At some point soon, I will be cleaning up the code and hopefully making it better!